check for 200 response code before parsing

+++ app/controllers/scrapes_controller.rb
@@ -46,37 +46,40 @@ class ScrapesController < ApplicationController
       agent = Mechanize.new
       agent.user_agent_alias = "Mac Safari"
       slink_doc = agent.get(slink)
-      slink_doc.xpath('//tr').each do |row|  # this is the loop for individual rows of data.
-        if row.children.length > 8  && row.children[0].children.text != '' && (!row.children[1].children.text.include? 'parkrunner')
-            time_in_seconds = row.children[2].children.text.split(':')[-1].to_i + row.children[2].children.text.split(':')[-2].to_i*60  + row.children[2].children.text.split(':')[-3].to_i*3600
-            result = Result.create(
-              pos:            row.children[0].children.text,
-              parkrunner:     row.children[1].children.text,
-              time:           time_in_seconds,
-              age_cat:        row.children[3].children.text,
-              age_grade:      row.children[4].children.text,
-              gender:         row.children[5].children.text,
-              gender_pos:     row.children[6].children.text,
-              club:           row.children[7].children.text,
-              note:           row.children[8].children.text,
-              total:          row.children[9].children.text,
-              run_id:         run_identifier.id,
-              athlete_number: get_runner_number_from_text(row) || nil
-              )
-            if ([49, 99, 149, 199, 249, 299, 349, 399, 449, 499, 549, 599, 649].include? result.total) && (run_identifier.run_identifier.include? 'astleigh' or result.club.include? 'astleigh')
-              milestone = Milestone.find_or_create_by(result.attributes.except('id', 'created_at', 'updated_at'))
-              Rails.logger.info "AWOOGA should create milestone, #{result.parkrunner}"
-            elsif  [9].include? result.total && (result.age_cat.include? 'J')
-              milestone=Milestone.find_or_create_by(result.attributes.except('id', 'created_at', 'updated_at'))
-            elsif ([10, 50, 100, 200, 250, 300, 350, 400, 450, 500, 550, 600, 650].include? result.total)
-              result_to_clear = Milestone.find_by athlete_number: result.athlete_number if Milestone.exists?(:athlete_number => result.athlete_number)
-              if(result_to_clear)
-                Rails.logger.debug "AWOOGA destroying milestone, #{result_to_clear.parkrunner}"
-                result_to_clear.destroy
+
+      if response.code == '200'
+        slink_doc.xpath('//tr').each do |row|  # this is the loop for individual rows of data.
+          if row.children.length > 8  && row.children[0].children.text != '' && (!row.children[1].children.text.include? 'parkrunner')
+              time_in_seconds = row.children[2].children.text.split(':')[-1].to_i + row.children[2].children.text.split(':')[-2].to_i*60  + row.children[2].children.text.split(':')[-3].to_i*3600
+              result = Result.create(
+                pos:            row.children[0].children.text,
+                parkrunner:     row.children[1].children.text,
+                time:           time_in_seconds,
+                age_cat:        row.children[3].children.text,
+                age_grade:      row.children[4].children.text,
+                gender:         row.children[5].children.text,
+                gender_pos:     row.children[6].children.text,
+                club:           row.children[7].children.text,
+                note:           row.children[8].children.text,
+                total:          row.children[9].children.text,
+                run_id:         run_identifier.id,
+                athlete_number: get_runner_number_from_text(row) || nil
+                )
+              if ([49, 99, 149, 199, 249, 299, 349, 399, 449, 499, 549, 599, 649].include? result.total) && (run_identifier.run_identifier.include? 'astleigh' or result.club.include? 'astleigh')
+                milestone = Milestone.find_or_create_by(result.attributes.except('id', 'created_at', 'updated_at'))
+                Rails.logger.info "AWOOGA should create milestone, #{result.parkrunner}"
+              elsif  [9].include? result.total && (result.age_cat.include? 'J')
+                milestone=Milestone.find_or_create_by(result.attributes.except('id', 'created_at', 'updated_at'))
+              elsif ([10, 50, 100, 200, 250, 300, 350, 400, 450, 500, 550, 600, 650].include? result.total)
+                result_to_clear = Milestone.find_by athlete_number: result.athlete_number if Milestone.exists?(:athlete_number => result.athlete_number)
+                if(result_to_clear)
+                  Rails.logger.debug "AWOOGA destroying milestone, #{result_to_clear.parkrunner}"
+                  result_to_clear.destroy
+                end
               end
-            end
-        end   # here ends each row operation
-      end # here ends the slink each
+          end   # here ends each row operation
+        end # here ends the slink each
+      end
     end  # here ends each link for scraping
 
     # now would be a good time to assign age grade positions.
